Split lsp/server.py: extract feature_config, state, settings, diagnostics_pipeline
Phases 1–4 of the lsp/server.py decomposition:

- lsp/feature_config.py: FeatureConfig dataclass (pure, no lsp.* deps)
- lsp/state.py: singleton state, process pool, utility helpers, semantic
  token cache, _extract_tcl_lsp_settings; imports codes_all before
  feature_config to ensure optimisation profile categories are fully
  populated at import time
- lsp/settings.py: _apply_feature_settings, debounce pipeline,
  _pull_and_apply_configuration, did_change_configuration registration
- lsp/diagnostics_pipeline.py: pull cache, push/pull handlers,
  _publish_diagnostics async pipeline, _publish_diagnostics_sync,
  dialect detection helpers, BIG-IP/APL publishers

server.py drops from 4122 → 2661 lines. All 66 affected tests pass.
