+++ tests/test_diagnostic_manifest.py
@@ -121,19 +121,19 @@ def test_internal_codes_are_real():
 
 def test_server_diagnostic_codes_match_registry():
     """The server's _ALL_DIAGNOSTIC_CODES matches the registry."""
-    import lsp.server as server_module
+    from lsp.settings import _ALL_DIAGNOSTIC_CODES
 
-    assert server_module._ALL_DIAGNOSTIC_CODES == diagnostic_codes(), (
-        "server._ALL_DIAGNOSTIC_CODES does not match registry"
+    assert _ALL_DIAGNOSTIC_CODES == diagnostic_codes(), (
+        "lsp.settings._ALL_DIAGNOSTIC_CODES does not match registry"
     )
 
 
 def test_server_optimisation_codes_match_registry():
     """The server's _ALL_OPTIMISATION_CODES matches the registry."""
-    import lsp.server as server_module
+    from lsp.settings import _ALL_OPTIMISATION_CODES
 
-    assert server_module._ALL_OPTIMISATION_CODES == optimisation_codes(), (
-        "server._ALL_OPTIMISATION_CODES does not match registry"
+    assert _ALL_OPTIMISATION_CODES == optimisation_codes(), (
+        "lsp.settings._ALL_OPTIMISATION_CODES does not match registry"
     )
